--- client/components/Tv.js
@@ -1,9 +1,0 @@
-import * as React from 'react'
-
-const TV = props => (
-  <div id="border">
-    <div className="tvlines">{props.children}</div>
-  </div>
-)
-
-export default TV
