+++ client/components/home/user-home.js
@@ -1,0 +1,36 @@
+import React from 'react'
+import PropTypes from 'prop-types'
+import {connect} from 'react-redux'
+import HomePage from './home-page'
+
+/**
+ * COMPONENT
+ */
+export const UserHome = props => {
+  const {email} = props
+
+  return (
+    <div>
+      <h3>Welcome, {email}</h3>
+      <HomePage />
+    </div>
+  )
+}
+
+/**
+ * CONTAINER
+ */
+const mapState = state => {
+  return {
+    email: state.user.email
+  }
+}
+
+export default connect(mapState)(UserHome)
+
+/**
+ * PROP TYPES
+ */
+UserHome.propTypes = {
+  email: PropTypes.string
+}
