+++ client/components/index.js
@@ -3,10 +3,16 @@
  * This way, we can import all of our components from the same place, rather than
  * having to figure out which file they belong to!
  */
-export {default as Navbar} from './navbar'
-export {default as UserHome} from './user-home'
-export {default as HomePage} from './home-page'
-export {default as SongMenu} from './song-menu'
-export {default as SelectedSongMenu} from './selected-song-menu'
-export {default as Modal} from './Modal'
-export {Login, Signup} from './auth-form'
+export {default as Navbar} from '../components/home/navbar'
+export {default as UserHome} from '../components/home/user-home'
+export {default as HomePage} from '../components/home/home-page'
+export {default as SongMenu} from '../components/menu/song-menu'
+export {
+  default as SelectedSongMenu
+} from '../components/menu/selected-song-menu'
+export {default as Modal} from '../components/menu/Modal'
+export {Login, Signup} from '../components/home/auth-form'
+export {default as Webcam} from '../components/main/webcam'
+export {default as Main} from '../components/main/Main'
+export {default as Sing} from '../components/main/Sing'
+export {default as score} from '../components/main/score'
