+++ client/components/main/Tv.js
@@ -1,0 +1,9 @@
+import * as React from 'react'
+
+const TV = props => (
+  <div id="border">
+    <div className="tvlines">{props.children}</div>
+  </div>
+)
+
+export default TV
